+++ docs/src/components/Sidebar/Sidebar.tsx
@@ -199,11 +199,17 @@ class Sidebar extends React.Component<any, any> {
             <Menu.Item as={NavLink} exact to="/quick-start" activeClassName="active">
               Quick Start
             </Menu.Item>
+            <Menu.Item as={NavLink} exact to="/glossary" activeClassName="active">
+              Glossary
+            </Menu.Item>
             <Menu.Item as={NavLink} exact to="/accessibility" activeClassName="active">
               Accessibility
             </Menu.Item>
             <Menu.Item as={NavLink} exact to="/theming" activeClassName="active">
               Theming
+            </Menu.Item>
+            <Menu.Item as={NavLink} exact to="/theming-examples" activeClassName="active">
+              Theming Examples
             </Menu.Item>
           </Menu.Menu>
         </Menu.Item>