+++ docs/src/examples/components/Text/Variations/TextExampleAtMention.shorthand.tsx
@@ -1,6 +1,12 @@
 import React from 'react'
 import { Text } from '@stardust-ui/react'
 
-const TextExampleAtMentionShorthand = () => <Text atMention content="@Russell Wilson" />
+const TextExampleAtMentionShorthand = () => (
+  <div>
+    <Text atMention content="@someone" />
+    <br />
+    <Text atMention="me" content="@me" />
+  </div>
+)
 
 export default TextExampleAtMentionShorthand