+++ src/components/Grid/Grid.tsx
@@ -1,9 +1,9 @@
 import * as PropTypes from 'prop-types'
 import * as React from 'react'
-import ReactNode = React.ReactNode
 import { UIComponent, childrenExist, customPropTypes, IRenderResultConfig } from '../../lib'
 import { ComponentVariablesInput, ComponentPartStyle } from '../../../types/theme'
 import { Extendable, ItemShorthand, ReactChildren } from '../../../types/utils'
+import ReactNode = React.ReactNode
 
 export interface IGridProps {
   as?: any
@@ -57,17 +57,6 @@ class Grid extends UIComponent<Extendable<IGridProps>, any> {
     /** Override for theme site variables to allow modifications of component styling via themes. */
     variables: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
   }
-
-  public static handledProps = [
-    'as',
-    'children',
-    'className',
-    'columns',
-    'content',
-    'rows',
-    'styles',
-    'variables',
-  ]
 
   public static defaultProps = {
     as: 'div',