+++ src/themes/teams/components/Chat/chatItemStyles.ts
@@ -1,0 +1,7 @@
+import { ICSSInJSStyle } from '../../../../../types/theme'
+
+const chatItemStyles = {
+  root: (): ICSSInJSStyle => ({}),
+}
+
+export default chatItemStyles